♻️ refactor formats crud

+++ app/Http/Controllers/FormatController.php
@@ -20,19 +20,22 @@ class FormatController extends Controller
     public function store(Request $request)
     {
         $fields = $request->validate([
-            'name'  => 'required|string|max:100|unique:formats',
-            'file'  => 'required|mimes:doc,docx,pdf',
+            'code'         => 'required|string|max:50|unique:formats',
+            'name'         => 'required|string|max:100',
+            'is_template'  => 'required|boolean',
+            'file'         => 'required|mimes:doc,docx,pdf',
         ]);
 
-        $formatName = preg_replace('/\s+/', '-', $fields['name']);
+        $formatName = preg_replace('/\s+/', '-', $fields['code']);
         $fileExt = pathinfo($fields['file']->getClientOriginalName(), PATHINFO_EXTENSION);
-        $fileUrl = $formatName.'.'.$fileExt;
-
-        \Storage::disk('formats')->put($fileUrl, \File::get($fields['file'])); 
+        $file_url = $formatName.'.'.$fileExt;
+        \Storage::disk('formats')->put($file_url, \File::get($fields['file'])); 
         
         $newFormat = Format::create([
-            'name'      => $fields['name'],
-            'fileUrl'   => $fileUrl,
+            'code'         => $fields['code'],
+            'name'         => $fields['name'],
+            'is_template'  => $fields['is_template'],
+            'file_url'     => $file_url,
         ]);
 
         $this->RegisterAction('El usuario ha ingresado un nuevo formato con id: '.$newFormat->id, 'medium');
@@ -42,34 +45,52 @@ class FormatController extends Controller
     public function show($id)
     {
         $format = Format::findOrFail($id);
-        $file = base64_encode(\Storage::disk('formats')->get($format->fileUrl));
+        $format->encoded_file = base64_encode(\Storage::disk('formats')->get($format->file_url));
         $this->RegisterAction('El usuario ha consultado el formato con id: '.$id);
-        return response(['encodedFile' => $file], 200);
+        return response($format, 200);
     }
      
     public function update(Request $request, $id)
     {
         $fields = $request->validate([
-            'name'  => 'required|string|max:100',
-            'file'  => 'required|mimes:doc,docx,pdf',
+            'code'        => 'required|string|max:100',
+            'name'        => 'required|string|max:100',
+            'is_template' => 'required|boolean',
+            'file'        => 'mimes:doc,docx,pdf',
         ]);
 
         $format = Format::findOrFail($id);
-        $formatByName = Format::withTrashed()->where('name', $fields['name'])->first();
-        if($formatByName && $formatByName->id != $format->id){
-            return response(['message' => 'El valor del campo name ya está en uso'], 400);
+
+        if($format->is_template && $fields['code'] != $format->code) {
+            return response(['message' => 'No se puede editar el codigo de un archivo plantilla'], 400);
         }
 
-        $formatName = preg_replace('/\s+/', '-', $fields['name']);
-        $fileExt = pathinfo($fields['file']->getClientOriginalName(), PATHINFO_EXTENSION);
-        $fileUrl = $formatName.'.'.$fileExt;
+        if($format->is_template == true && $fields['is_template'] != true) {
+            return response(['message' => 'No se puede eliminar la etiqueta de plantilla de un archivo'], 400);
+        }
 
-        \Storage::disk('formats')->delete($format->fileUrl);
-        \Storage::disk('formats')->put($fileUrl, \File::get($fields['file'])); 
+        if($fields['code'] != null) {
+            $formatByCode = Format::withTrashed()->where('code', $fields['code'])->first();
+            if($formatByCode && $formatByCode->id != $format->id) {
+                return response(['message' => 'El valor del campo código ya ha sido ocupado'], 400);
+            }
+        }
+
+        if($fields['file']){
+            $formatName = preg_replace('/\s+/', '-', $fields['code']);
+            $fileExt = pathinfo($fields['file']->getClientOriginalName(), PATHINFO_EXTENSION);
+            $file_url = $formatName.'.'.$fileExt;
+    
+            \Storage::disk('formats')->delete($format->file_url);
+            \Storage::disk('formats')->put($file_url, \File::get($fields['file'])); 
+        } else {
+            $file_url = $format->file_url;
+        }
+
 
        $format->update([
             'name'      => $fields['name'],
-            'fileUrl'   => $fileUrl,
+            'file_url'  => $file_url,
         ]);
 
         $this->RegisterAction('El usuario ha actualizado el formato con id: '.$id, 'medium');
@@ -79,8 +100,11 @@ class FormatController extends Controller
     public function destroy($id)
     {
         $format = Format::findOrFail($id);
+        if($format->is_template){
+            return response(['message' => 'No se pueden eliminar los archivos plantilla'], 400);
+        }
         $format->delete();
-        \Storage::disk('formats')->delete($format->fileUrl);
+        \Storage::disk('formats')->delete($format->file_url);
 
         $this->RegisterAction('El usuario ha eliminado el formato con id: '.$id, 'medium');
         return response(null, 204);
